+++ src/main/java/pl/gameChecker/model/UsersLogin.java
@@ -1,0 +1,37 @@
+package pl.gameChecker.model;
+
+/**
+ *
+ * @author mlethys
+ */
+public class UsersLogin {
+    
+    private String username;
+    private String password;
+    
+    public UsersLogin() {
+        this.username = "";
+        this.password = "";
+    }
+    
+    public UsersLogin(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+    
+    public String getUsername() {
+        return this.username;
+    }
+    
+    public String getPassword() {
+        return this.password;
+    }
+    
+    public void setUsername(String username) {
+        this.username = username;
+    }
+    
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
