+++ src/main/java/pl/gameChecker/model/hibernateEntities/Gametype.java
@@ -4,7 +4,8 @@
  * and open the template in the editor.
  */
 
-package com.bbZoftware.hibernateEntities;
+package pl.gameChecker.model.hibernateEntities;
+
 
 import java.io.Serializable;
 import java.util.List;
