+++ src/main/java/pl/gameChecker/model/hibernateEntities/Library.java
@@ -1,10 +1,11 @@
-package com.bbZoftware.hibernateEntities;
+package pl.gameChecker.model.hibernateEntities;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+
 
 
 import java.io.Serializable;
