--- src/main/webapp/WEB-INF/mvc-dispatcher-servlet.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="
-        http://www.springframework.org/schema/beans     
-        http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-        http://www.springframework.org/schema/context 
-        http://www.springframework.org/schema/context/spring-context-3.0.xsd">
- 
-	<context:component-scan base-package="pl.mlethys.java.gamecheckerservice.controller" />
- 
-	<bean class="org.springframework.web.servlet.view.InternalResourceViewResolver">
-            <property name="prefix">
-                <value>/</value>
-            </property>
-            <property name="suffix">
-                <value>.html</value>
-            </property>
-        </bean>
-</beans>
