+++ src/main/webapp/index.html
@@ -40,30 +40,8 @@
             </div>
             <div id="mainBody">
                 <div id="newsContainer" class="mainBody">
-                    <h1 class="headline">Second news!</h1>
-                    <p class="newsText">This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text.</p>
-                    <h1 class="headline">First news!</h1>
-                    <p class="newsText">This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text. 
-                    This is some text. This is some text. This is some text.</p>
+                    <h1 class="headline">${newsTitle}</h1>
+                    <p class="newsText">${newsContent}</p>
                 </div>
                 <div id="addsContainer" class="mainBody">
                     <img src="images/sampleAdd.jpg" alt="add" class="add"/>
