+++ src/main/webapp/login.html
@@ -9,7 +9,7 @@
     <body>
         <div id="loginContainer" class="loginRegister" align="center">
             <h1 class="logRegHeadline">Log in</h1>
-            <form id="loginForm" method="POST" name="login">
+            <form id="loginForm" method="POST" name="login" action="Login">
                 <p class="loginRegisterText">Login</p>
                 <input type="text" name="login"/>
                 <br>
